explorer parsing from files array
