+++ src/components/plugin/pluginOutput.tsx
@@ -20,7 +20,7 @@ const PluginOutput: React.FunctionComponent<AllProps> = (props: AllProps) => {
   };
 
   // Set local state hook
-  const [isModalOpen, setValue] = useState(false);
+  const [isModalOpen, setValue] = useState(false); // Temp - set to false
   const handleModalToggle = () => {
     setValue(!isModalOpen);
   };
