+++ src/store/plugin/actions.ts
@@ -1,18 +1,15 @@
-
 import { action } from "typesafe-actions";
 import { PluginActionTypes } from "./types";
 import { IPluginItem } from "../../api/models/pluginInstance.model";
-
+import UITreeNodeModel, { IUITreeNode } from "../../api/models/file-explorer";
 
 export const getPluginDescendantsRequest = (url: string) => action(PluginActionTypes.GET_PLUGIN_DESCENDANTS, url);
 export const getPluginDescendantsSuccess = (items: IPluginItem[]) => action(PluginActionTypes.GET_PLUGIN_DESCENDANTS_SUCCESS, items);
 
-export const getPluginFilesRequest = (url: string) => action(PluginActionTypes.GET_PLUGIN_FILES, url);
+export const getPluginFilesRequest = (url: string) =>  action(PluginActionTypes.GET_PLUGIN_FILES, url);
 export const getPluginFilesSuccess = (items: IPluginItem[]) => action(PluginActionTypes.GET_PLUGIN_FILES_SUCCESS, items);
-export const setExplorerSuccess = (items: IPluginItem[]) => {
-    console.log("setExplorerSuccess", items);
-    return  action(PluginActionTypes.SET_EXPLORER_SUCCESS, items);
-};
+// Description: Parse the files array into a File tree obj
+export const setExplorerSuccess = (items: IPluginItem[]) => action(PluginActionTypes.SET_EXPLORER_SUCCESS, UITreeNodeModel.parseUiTree(items));
 
 
 export const getPluginParametersRequest = (url: string) => action(PluginActionTypes.GET_PLUGIN_PARAMETERS, url);
@@ -20,4 +17,3 @@ export const getPluginParametersSuccess = (items: IPluginItem[]) => action(Plugi
 
 export const getPluginDetailsRequest = (item: IPluginItem) => action(PluginActionTypes.GET_PLUGIN_DETAILS, item);
 export const getPluginDetailsSuccess = (items: IPluginItem[]) => action(PluginActionTypes.GET_PLUGIN_DETAILS_SUCCESS, items);
-
